separate armv7 assembler
